+++ bpam/apps/common/admin.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import logging
 from functools import partial
 from django.contrib import admin
 from django import forms
@@ -25,7 +26,8 @@ from .models import SampleSite
 
 JIRA_URL = "https://ccgmurdoch.atlassian.net/projects/BRLOPS/issues/"
 
-DEGREES = u'°'
+
+logger = logging.getLogger(__name__)
 
 
 class BPAImportExportModelAdmin(ImportExportModelAdmin):
@@ -52,37 +54,18 @@ class BPAModelResource(resources.ModelResource):
 
     def transform_row(self, row):
         transformations = {}
-
-        # These look like common enough problems to have them in the base class
-
-        # 102.100.100/34956 -> 34956
-        if '/' in row.get('BPA_ID', ''):
-            bpa_id = row.get('BPA_ID').split('/')[-1]
-            if isinteger(bpa_id):
-                transformations['BPA_ID'] = bpa_id
-
-        # accepts time both with ('%H:%M') and without seconds ('%H:%M:%S')
-        time_sampled = row.get('Time Sampled', '')
-        if isshorttime(time_sampled):
-            if istime('%s:00' % time_sampled):
-                transformations['Time Sampled'] = '%s:00' % time_sampled
-
-        # removes possible DEGREES character from Longitude
-        if row.get('Longitude', '').rstrip().endswith(DEGREES):
-            transformations['Longitude'] = row.get('Longitude').rstrip().rstrip(DEGREES)
-
-        # removes possible DEGREES character from Latitude
-        if row.get('Latitude', '').rstrip().endswith(DEGREES):
-            transformations['Latitude'] = row.get('Latitude').rstrip().rstrip(DEGREES)
-
-
+        # Override and add your transformations here
         return transformations
 
     def maybe_transform_row(self, row):
         transformations = self.transform_row(row)
 
         if len(transformations) > 0:
+            logger.info('Transforming row')
             new_row = row.copy()
+            # log the column headers in the order of the row (OrderedDict)
+            for name in filter(lambda x: x in transformations, row):
+                logger.info("'%s': '%s' -> '%s'", name, row.get(name), transformations.get(name))
             new_row.update(transformations)
             return new_row
         return row
